Fix GameData unkC24 signedness

+++ src/music.c
@@ -158,7 +158,7 @@ void sub_804AE34(unk32 arg0, unk32 arg1)
 
 void sub_804AE8C(void)
 {
-    if ((s16)_gameData->unkC24 != -1) {
+    if (_gameData->unkC24 != -1) {
         Sound_8062A90(_gameData->unkC28);
     }
     _gameData->unkC24 = -1;
@@ -166,7 +166,7 @@ void sub_804AE8C(void)
 
 void sub_804AECC(unk32 arg0)
 {
-    if ((s16)_gameData->unkC24 != -1) {
+    if (_gameData->unkC24 != -1) {
         Sound_8062AF4(_gameData->unkC28, arg0);
     }
 }
